add updateCacheOnForeclosure behaviour to MintTicketRow
